pyproject.toml
[project]
name = "book-list"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datetime>=5.5",
    "jaconv>=0.4.0",
    "japanize-matplotlib>=1.1.3",
    "logging>=0.4.9.6",
    "matplotlib>=3.10.6",
    "mysql>=0.0.3",
    "mysql-connector-python>=9.4.0",
    "mysqlclient>=2.2.7",
    "pandas>=2.3.2",
    "regex>=2025.9.1",
    "sqlalchemy>=2.0.43",
    "typing>=3.10.0.0",
]
